first set of modifications done BUT NEED TO VERIFY git add -u

+++ include/MCclassic.h
@@ -1,7 +1,7 @@
 #ifndef MCCLASSIC_H_INCLUDED
 #define MCCLASSIC_H_INCLUDED
 
-int make_MC_moves(ATOM at[], DATA *dat, double *ener);
-int apply_Metrop(ATOM at[], ATOM at_new[], DATA *dat, int *candidate, double *ener, int *step);
+uint64_t make_MC_moves(ATOM at[], DATA *dat, double *ener);
+int32_t apply_Metrop(ATOM at[], ATOM at_new[], DATA *dat, int32_t *candidate, double *ener, uint64_t *step);
 
 #endif // MCCLASSIC_H_INCLUDED
